Move custom slider to theme
