+++ lib/app/router.dart
@@ -5,6 +5,7 @@ import 'package:mistral_ai_chat_example_app/app/home_page.dart';
 import 'package:mistral_ai_chat_example_app/mistral_ai_book_search_example/mistralai_book_search_page.dart';
 import 'package:mistral_ai_chat_example_app/mistral_ai_chat_example/mistralai_chat_page.dart';
 import 'package:mistral_ai_chat_example_app/mistral_ai_llm_controller_example/mistral_ai_llm_controller_page.dart';
+import 'package:mistral_ai_chat_example_app/mistral_ai_summary_example/mistralai_summary_page.dart';
 
 part 'router.g.dart';
 
@@ -40,7 +41,7 @@ class MistralAISummaryRoute extends GoRouteData {
 
   @override
   Page<void> buildPage(BuildContext context, GoRouterState state) {
-    return _buildPage(const MistralAIBookSearchPage());
+    return _buildPage(const MistralAISummaryPage());
   }
 }
